Condensed duplicated code and simplified it
